requirements.txt
pytest
pytest-cov
datetime
configparser
pydantic
azure-functions
azure.identity==1.6.0
azure.keyvault.secrets==4.2.0
azure.storage.blob==12.8.1
cryptography==3.4.7
pandas==1.3.5
azure-storage-file-datalake
cffi==1.14.6
black
mypy==0.910
interrogate==1.5.0
pyodbc
pyapacheatlas